requirements.txt
asgiref==3.11.1
Django==4.2.28
django-modeltranslation==0.19.17
pillow==11.3.0
psycopg2-binary==2.9.11
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
